backend/resume_ai/apps/job_postings/serializers.py
from rest_framework import serializers
from .models import JobPosting

class JobPostingSerializer(serializers.ModelSerializer):
    class Meta:
        model = JobPosting
        fields = ['id', 'title','company', 'description', 'created_at', 'updated_at']